--- numerous/declarative/connectors/exceptions.py
@@ -1,11 +1,0 @@
-class ChannelAlreadyOnBusError(Exception):
-    ...
-
-class AlreadyConnectedError(Exception):
-    ...
-
-class ChannelNotFound(Exception):
-    ...
-
-class WrongMappingDirectionError(Exception):
-    ...
