+++ tests/test_declarative/test_module.py
@@ -1,4 +1,6 @@
-from numerous.declarative.specification import create_mappings, ItemsSpec, MappingOutsideMappingContextError, Module, ScopeSpec, EquationSpec, NotMappedError, FixedMappedError
+from numerous.declarative.specification import ItemsSpec, Module, ScopeSpec, EquationSpec
+from numerous.declarative.mappings import create_mappings
+from numerous.declarative.exceptions import MappingOutsideMappingContextError, NotMappedError, FixedMappedError
 from numerous.declarative.variables import Parameter, Constant
 
 import pytest
